+++ src/main/java/kr/hhplus/be/server/domain/order/OrderEvent.java
@@ -13,4 +13,10 @@ public record OrderEvent(
         OrderStatus status,
         LocalDateTime changedAt
 ) {
+    public enum OrderStatus {
+        CREATED,
+        PAID,
+        CANCELED,
+        COMPLETED
+    }
 }
