+++ src/main/java/kr/hhplus/be/server/domain/order/OrderItem.java
@@ -4,6 +4,7 @@ package kr.hhplus.be.server.domain.order;
 import lombok.Builder;
 
 import java.math.BigDecimal;
+import java.util.List;
 
 @Builder
 public record OrderItem(
@@ -16,5 +17,11 @@ public record OrderItem(
 ) {
     public static BigDecimal calculateTotal(BigDecimal price, long quantity) {
         return price.multiply(BigDecimal.valueOf(quantity));
+    }
+
+    public static BigDecimal sumTotalPrice(List<OrderItem> items) {
+        return items.stream()
+                .map(OrderItem::totalPrice)
+                .reduce(BigDecimal.ZERO, BigDecimal::add);
     }
 }
