--- src/main/java/kr/hhplus/be/server/domain/order/OrderStatus.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.domain.order;
-
-public enum OrderStatus {
-    CREATED,
-    PAID,
-    CANCELED,
-    COMPLETED
-}
